pod-scaler: factor out types for reuse
We need these in tests.
